move prepare data from plan to test
